refactor: move restriction components to subfolder
