+++ dev-client/src/screens/LocationScreens/components/soilInfo/SoilIdMatchSelector.tsx
@@ -23,7 +23,7 @@ import {DataBasedSoilMatch} from 'terraso-client-shared/graphqlSchema/graphql';
 
 import {DisableableText} from 'terraso-mobile-client/components/content/typography/DisableableText';
 import {TranslatedParagraph} from 'terraso-mobile-client/components/content/typography/TranslatedParagraph';
-import {RestrictBySiteRole} from 'terraso-mobile-client/components/RestrictByRole';
+import {RestrictBySiteRole} from 'terraso-mobile-client/components/restrictions/RestrictByRole';
 import {useIsOffline} from 'terraso-mobile-client/hooks/connectivityHooks';
 import {SITE_EDITOR_ROLES} from 'terraso-mobile-client/model/permissions/permissions';
 import {getMatchSelectionId} from 'terraso-mobile-client/model/soilId/soilMetadataFunctions';
