+++ dev-client/src/screens/ProjectTeamScreen/ProjectTeamScreen.tsx
@@ -28,7 +28,7 @@ import {
   Column,
   Heading,
 } from 'terraso-mobile-client/components/NativeBaseAdapters';
-import {RestrictByProjectRole} from 'terraso-mobile-client/components/RestrictByRole';
+import {RestrictByProjectRole} from 'terraso-mobile-client/components/restrictions/RestrictByRole';
 import {useProjectRoleContext} from 'terraso-mobile-client/context/ProjectRoleContext';
 import {PROJECT_MANAGER_ROLES} from 'terraso-mobile-client/model/permissions/permissions';
 import {deleteUserFromProject} from 'terraso-mobile-client/model/project/projectSlice';
