+++ dev-client/src/screens/SoilScreen/ColorScreen/ColorScreen.tsx
@@ -31,7 +31,7 @@ import {
   Row,
   Text,
 } from 'terraso-mobile-client/components/NativeBaseAdapters';
-import {RestrictBySiteRole} from 'terraso-mobile-client/components/RestrictByRole';
+import {RestrictBySiteRole} from 'terraso-mobile-client/components/restrictions/RestrictByRole';
 import {SiteRoleContextProvider} from 'terraso-mobile-client/context/SiteRoleContext';
 import {isColorComplete} from 'terraso-mobile-client/model/color/colorConversions';
 import {MunsellColor} from 'terraso-mobile-client/model/color/types';
